+++ .php-cs-fixer.dist.php
@@ -10,14 +10,12 @@ For the full copyright and license information, please view the LICENSE
 file that was distributed with this source code.
 EOF;
 
-return PhpCsFixer\Config::create()
+return (new PhpCsFixer\Config())
     ->setRules([
-        '@PSR2' => true,
-        'array_syntax' => ['syntax' => 'short'],
+        '@PSR12:risky' => true,
         'header_comment' => [
             'header' => $header,
         ],
-        'ordered_imports' => true,
     ])
     ->setLineEnding("\n")
     ->setUsingCache(false)